Prevent email attachments from not showing up on Android M or higher
Because Android M and higher adds a new permission model, it requires you to start
using FileProviders. When you don't use them, depending on your Email client it will
ask you for specific permissions, or just not work at all.

It first checks if there are attachments to add, and if ran on Android M or higher outputs
a log warning that mentions the developer should use FileProviders.

This solution requires the use of Android Support Library v4 (FileProvider). That's
the reason this reference is added.

fixes #1110